Two different 'typeof' format and 'BigInt' data type
In JavaScript, the “number” type cannot represent integer values larger than (2^53-1) (that’s 9007199254740991), or less than -(2^53-1) for negatives. Sometimes we need really large numbers like in cryptography etc. So, to remove this limitation BigInt was added to the language to represent integers of any length. 
To represent BigInt we just need to put n at the end of the integer